firmware: Add code to use I2CA as a master
This works well enough to read/write registers from the charger IC.

The code doesn't work for passing the data back to the host (via I2CB)
because it is interrupt driven, but is already called from an interrupt
handler. Using the third bank of registers (for the nested ISR) breaks
the firmware, most likely because the expected location of the stack is
moved (the firmware does NOT currently initialize the stack pointer!).